requirements.txt
av==10.0.0
decord==0.6.0
einops==0.6.1
imageio==2.27.0
imageio-ffmpeg==0.4.8
kornia
matplotlib==3.9.2
moviepy==1.0.3
numpy==1.26.0
open-clip-torch==2.17.1
opencv-python==4.7.0.72
Pillow==9.4.0
pip==23.0.1
pyglet==1.5.0
pytorch-lightning==1.9.3
PyYAML==6.0
roma==1.5.0
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.9.1
tensorboard==2.12.2
timm==0.6.13
tqdm==4.65.0
transformers==4.28.1
trimesh==4.4.3
xformers==0.0.23.post1
gradio==3.37.0
gradio_client==0.7.1
omegaconf==2.3.0
triton
altair==5.4.0
certifi==2024.7.4
grpcio==1.66.0
httpx==0.27.0
Pygments==2.18.0
starlette==0.38.2
tifffile==2024.8.24
yarl==1.9.4
huggingface-hub==0.25.2
ninja==1.11.1.1
pandas==2.2.2
safetensors==0.4.3
wandb==0.17.4
neptune==1.12.0
plyfile==1.1
torchmetrics==1.5.0
hydra-core==1.3.2
jaxtyping==0.2.34
GitPython==3.1.43